Make MIR cleanup for functions with impossible predicates into a real MIR pass
